bug fix: PLCAPI (e.g. DeletetPerson) delete an object and then call registry.remove, which again tries to remove the same object using PLCAPI call

+++ sfa/methods/remove.py
@@ -43,6 +43,7 @@ class remove(Method):
         type = record['type']
         # TODO: sa, ma
         if type == "user":
+	  if self.api.plshell.GetPersons(self.api.plauth, record.get_pointer()):
             self.api.plshell.DeletePerson(self.api.plauth, record.get_pointer())
         elif type == "slice":
             self.api.plshell.DeleteSlice(self.api.plauth, record.get_pointer())
